+++ src/extension.ts
@@ -14,13 +14,12 @@ export function activate(context: vscode.ExtensionContext) {
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-	let disposable = vscode.commands.registerCommand('btrfsFileHistory.getHistory', async (contextSelection: vscode.Uri) => {
-
-		await getHistory(contextSelection);
-
+	let getHisoryCommand = vscode.commands.registerCommand('btrfsFileHistory.getHistory', async (contextSelection: vscode.Uri) => 
+	{
+		await getHistory(contextSelection, context);
 	});
 
-	context.subscriptions.push(disposable);
+	context.subscriptions.push(getHisoryCommand);
 }
 
 // This method is called when your extension is deactivated