+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "btrfs-file-history" extension will be documented in this file.
 
+## 1.1.2
+
+- Changed the snapshot history results from a quick pick menu to a fully-featured webview.
+
 ## 1.0.2
 
-- Initial release
+- Initial release of BTRFS File History.